+++ app/src/main/java/org/sopt/certi/presentation/ui/activity/sideEffect/ActivitySideEffect.kt
@@ -1,0 +1,5 @@
+package org.sopt.certi.presentation.ui.activity.sideEffect
+
+sealed interface ActivitySideEffect {
+    data object showDeleteDialog : ActivitySideEffect
+}
