+++ app/src/main/java/org/sopt/certi/presentation/ui/main/MainNavHost.kt
@@ -9,16 +9,19 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.compose.NavHost
+import org.sopt.certi.presentation.ui.activity.navigation.activityNavGraph
 import org.sopt.certi.presentation.ui.certdetail.navigation.certDetailNavGraph
 import org.sopt.certi.presentation.ui.certlist.navigation.certListNavGraph
 import org.sopt.certi.presentation.ui.certrecommend.navigation.certRecommendNavGraph
 import org.sopt.certi.presentation.ui.home.navigation.homeNavGraph
 import org.sopt.certi.presentation.ui.login.navigation.loginNavGraph
+import org.sopt.certi.presentation.ui.myCert.navigation.myCertNavGraph
 import org.sopt.certi.presentation.ui.onboarding.OnBoardingViewModel
 import org.sopt.certi.presentation.ui.onboarding.navigation.onBoardingNavGraph
 import org.sopt.certi.presentation.ui.resume.navigation.resumeNavGraph
 import org.sopt.certi.presentation.ui.search.navigation.searchNavGraph
 import org.sopt.certi.presentation.ui.splash.navigation.splashNavGraph
+import org.sopt.certi.presentation.ui.workExperience.navigation.workExperienceNavGraph
 
 @Composable
 fun MainNavHost(
@@ -82,6 +85,21 @@ fun MainNavHost(
             )
 
             resumeNavGraph(
+                padding = padding,
+                navController = navigator.navController
+            )
+
+            myCertNavGraph(
+                padding = padding,
+                navController = navigator.navController
+            )
+
+            activityNavGraph(
+                padding = padding,
+                navController = navigator.navController
+            )
+
+            workExperienceNavGraph(
                 padding = padding,
                 navController = navigator.navController
             )
