+++ app/src/main/java/org/sopt/certi/presentation/ui/myCert/sideEffect/MyCertSideEffect.kt
@@ -1,0 +1,5 @@
+package org.sopt.certi.presentation.ui.myCert.sideEffect
+
+sealed interface MyCertSideEffect {
+    data object ShowDeleteDialog : MyCertSideEffect
+}
