+++ app/src/main/java/org/sopt/certi/presentation/ui/myCert/state/MyCertUiState.kt
@@ -1,0 +1,17 @@
+package org.sopt.certi.presentation.ui.myCert.state
+
+import org.sopt.certi.core.state.UiState
+import org.sopt.certi.domain.model.certification.CertificationData
+
+data class MyCertUiState(
+    val myCertListLoadState: UiState<List<CertificationData>>,
+    val selectedCertificationId: Long?
+) {
+    val loadState: UiState<Unit>
+        get() = when (myCertListLoadState) {
+            is UiState.Success -> UiState.Success(Unit)
+            is UiState.Failure -> UiState.Failure("fail to load data")
+            is UiState.Loading -> UiState.Loading
+            else -> UiState.Empty
+        }
+}
