--- app/src/main/java/org/sopt/certi/presentation/ui/resume/activity/sideEffect/ActivitySideEffect.kt
@@ -1,5 +1,0 @@
-package org.sopt.certi.presentation.ui.resume.activity.sideEffect
-
-sealed interface ActivitySideEffect {
-    data object showDeleteDialog : ActivitySideEffect
-}
