--- app/src/main/java/org/sopt/certi/presentation/ui/resume/activity/state/AddActivityUiState.kt
@@ -1,10 +1,0 @@
-package org.sopt.certi.presentation.ui.resume.activity.state
-
-data class AddActivityUiState(
-    val startDate: String,
-    val endDate: String,
-    val organizationValue: String,
-    val activityValue: String,
-    val descriptionValue: String,
-    val addButtonEnabled: Boolean
-)
