--- app/src/main/java/org/sopt/certi/presentation/ui/resume/main/sideEffect/ResumeSideEffect.kt
@@ -1,5 +1,0 @@
-package org.sopt.certi.presentation.ui.resume.main.sideEffect
-
-sealed interface ResumeSideEffect {
-    data object ShowCertificationDetailModal : ResumeSideEffect
-}
