--- app/src/main/java/org/sopt/certi/presentation/ui/resume/myCert/sideEffect/MyCertSideEffect.kt
@@ -1,5 +1,0 @@
-package org.sopt.certi.presentation.ui.resume.myCert.sideEffect
-
-sealed interface MyCertSideEffect {
-    data object ShowDeleteDialog : MyCertSideEffect
-}
