+++ app/src/main/java/org/sopt/certi/presentation/ui/resume/navigation/ResumeNavigation.kt
@@ -5,15 +5,9 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
-import androidx.navigation.navOptions
 import org.sopt.certi.core.navigation.MainTabRoute
 import org.sopt.certi.core.navigation.ResumeRoute
-import org.sopt.certi.presentation.ui.resume.activity.ResumeActivitiesRoute
-import org.sopt.certi.presentation.ui.resume.workExperience.ResumeAddWorkExperienceRoute
-import org.sopt.certi.presentation.ui.resume.activity.ResumeAddActivitiesRoute
-import org.sopt.certi.presentation.ui.resume.myCert.ResumeMyCertRoute
-import org.sopt.certi.presentation.ui.resume.main.ResumeRoute
-import org.sopt.certi.presentation.ui.resume.workExperience.ResumeWorkExperienceRoute
+import org.sopt.certi.presentation.ui.resume.ResumeRoute
 
 fun NavController.navigateToResume(navOptions: NavOptions) {
     navigate(MainTabRoute.Resume, navOptions)
@@ -27,16 +21,8 @@ fun NavController.navigateToWorkExperience() {
     navigate(ResumeRoute.WorkExperience)
 }
 
-fun NavController.navigateToAddWorkExperience() {
-    navigate(ResumeRoute.AddWorkExperience)
-}
-
 fun NavController.navigateToActivities() {
     navigate(ResumeRoute.Activities)
-}
-
-fun NavController.navigateToAddActivities() {
-    navigate(ResumeRoute.AddActivities)
 }
 
 fun NavGraphBuilder.resumeNavGraph(
@@ -49,50 +35,6 @@ fun NavGraphBuilder.resumeNavGraph(
             navigateToMyCert = navController::navigateToMyCert,
             navigateToWorkExperience = navController::navigateToWorkExperience,
             navigateToActivities = navController::navigateToActivities
-        )
-    }
-
-    composable<ResumeRoute.MyCert> {
-        ResumeMyCertRoute(padding = padding)
-    }
-
-    composable<ResumeRoute.WorkExperience> {
-        ResumeWorkExperienceRoute(
-            padding = padding,
-            onNavigateToAddWordExperience = navController::navigateToAddWorkExperience
-        )
-    }
-
-    composable<ResumeRoute.AddWorkExperience> {
-        ResumeAddWorkExperienceRoute(
-            padding = padding,
-            onNavigateToResume = {
-                navController.navigateToResume(
-                    navOptions {
-                        popUpTo(MainTabRoute.Resume) { inclusive = true }
-                    }
-                )
-            }
-        )
-    }
-
-    composable<ResumeRoute.Activities> {
-        ResumeActivitiesRoute(
-            padding = padding,
-            onNavigateToAddActivities = navController::navigateToAddActivities
-        )
-    }
-
-    composable<ResumeRoute.AddActivities> {
-        ResumeAddActivitiesRoute(
-            padding = padding,
-            navigateToResume = {
-                navController.navigateToResume(
-                    navOptions {
-                        popUpTo(MainTabRoute.Resume) { inclusive = true }
-                    }
-                )
-            }
         )
     }
 }
