+++ app/src/main/java/org/sopt/certi/presentation/ui/resume/sideEffect/ResumeSideEffect.kt
@@ -1,0 +1,5 @@
+package org.sopt.certi.presentation.ui.resume.sideEffect
+
+sealed interface ResumeSideEffect {
+    data object ShowCertificationDetailModal : ResumeSideEffect
+}
