--- app/src/main/java/org/sopt/certi/presentation/ui/resume/workExperience/ResumeWorkExperienceScreen.kt
@@ -1,120 +1,0 @@
-package org.sopt.certi.presentation.ui.resume.workExperience
-
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.compose.LocalLifecycleOwner
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.lifecycle.flowWithLifecycle
-import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.toImmutableList
-import org.sopt.certi.R
-import org.sopt.certi.core.component.dialog.CertiDeleteDialog
-import org.sopt.certi.core.state.UiState
-import org.sopt.certi.core.util.screenHeightDp
-import org.sopt.certi.core.util.screenWidthDp
-import org.sopt.certi.domain.model.ActivityData
-import org.sopt.certi.presentation.ui.resume.component.ResumeAddButton
-import org.sopt.certi.presentation.ui.resume.component.ResumeEditListItem
-import org.sopt.certi.presentation.ui.resume.workExperience.sideEffect.WorkExperienceSideEffect
-import org.sopt.certi.ui.theme.CertiTheme
-
-@Composable
-fun ResumeWorkExperienceRoute(
-    padding: PaddingValues,
-    onNavigateToAddWordExperience: () -> Unit,
-    viewModel: WorkExperienceViewModel = hiltViewModel()
-) {
-    val uiState by viewModel.workExperienceUiState.collectAsStateWithLifecycle()
-    val lifecycleOwner = LocalLifecycleOwner.current
-    var onDeleteDialogShow by remember { mutableStateOf(false) }
-
-    LaunchedEffect(Unit) {
-        viewModel.getWorkExperienceList()
-    }
-
-    LaunchedEffect(viewModel.sideEffect, lifecycleOwner) {
-        viewModel.sideEffect.flowWithLifecycle(lifecycleOwner.lifecycle).collect {
-            when (it) {
-                WorkExperienceSideEffect.showDeleteDialog -> onDeleteDialogShow = true
-            }
-        }
-    }
-
-    when (uiState.loadState) {
-        is UiState.Success -> ResumeWorkExperienceScreen(
-            onNavigateToAddWordExperience = onNavigateToAddWordExperience,
-            resumeDataList = (uiState.experienceListLoadState as UiState.Success<List<ActivityData>>).data.toImmutableList(),
-            onDeleteClick = { viewModel.onDeleteClick(it) },
-            modifier = Modifier.padding(padding)
-        )
-
-        is UiState.Empty -> {}
-        is UiState.Failure -> {}
-        is UiState.Init -> {}
-        is UiState.Loading -> {}
-    }
-
-    if (onDeleteDialogShow) {
-        CertiDeleteDialog(
-            onConfirmClick = {
-                onDeleteDialogShow = false
-                viewModel.onDeleteConfirmclick()
-            },
-            onDismissClick = { onDeleteDialogShow = false }
-        )
-    }
-}
-
-@Composable
-fun ResumeWorkExperienceScreen(
-    onNavigateToAddWordExperience: () -> Unit,
-    resumeDataList: ImmutableList<ActivityData>,
-    onDeleteClick: (Long) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    LazyColumn(
-        modifier = modifier.fillMaxSize(),
-        contentPadding = PaddingValues(horizontal = screenWidthDp(20.dp))
-    ) {
-        item {
-            ResumeAddButton(
-                onClick = onNavigateToAddWordExperience,
-                modifier = Modifier.padding(
-                    top = screenHeightDp(60.dp),
-                    bottom = screenHeightDp(48.dp)
-                )
-            )
-        }
-
-        item {
-            Text(
-                text = stringResource(R.string.resume_work_experience_edit_title),
-                style = CertiTheme.typography.subtitle.semibold_20,
-                color = CertiTheme.colors.gray600,
-                modifier = Modifier.padding(bottom = screenHeightDp(48.dp))
-            )
-        }
-
-        items(resumeDataList) { resumeData ->
-            ResumeEditListItem(
-                resumeListItem = resumeData,
-                onDeleteClick = onDeleteClick,
-                modifier = Modifier.padding(bottom = screenHeightDp(36.dp))
-            )
-        }
-    }
-}
