--- app/src/main/java/org/sopt/certi/presentation/ui/resume/workExperience/sideEffect/WorkExperienceSideEffect.kt
@@ -1,5 +1,0 @@
-package org.sopt.certi.presentation.ui.resume.workExperience.sideEffect
-
-sealed interface WorkExperienceSideEffect {
-    data object showDeleteDialog : WorkExperienceSideEffect
-}
