--- app/src/main/java/org/sopt/certi/presentation/ui/resume/workExperience/state/AddWorkExperienceUiState.kt
@@ -1,10 +1,0 @@
-package org.sopt.certi.presentation.ui.resume.workExperience.state
-
-data class AddWorkExperienceUiState(
-    val startDate: String,
-    val endDate: String,
-    val organizationValue: String,
-    val roleValue: String,
-    val descriptionValue: String,
-    val addButtonEnabled: Boolean
-)
