+++ app/src/main/java/org/sopt/certi/presentation/ui/workExperience/sideEffect/WorkExperienceSideEffect.kt
@@ -1,0 +1,5 @@
+package org.sopt.certi.presentation.ui.workExperience.sideEffect
+
+sealed interface WorkExperienceSideEffect {
+    data object showDeleteDialog : WorkExperienceSideEffect
+}
